Slate: STileView: STableViewBase: Fixed an issue where fully visible STileViews would inappropriately consume mouse wheel messages. The fix was done in two parts. First there was a floating point precision issue with determining if we are at the end of the list. Added IsNearlyZero() to fix this. The second issue was that the stats returned from RegenerateWidgets in the STileView could not return expected results. This was because the logic was assuming that ExactNumWidgetsOnScreen returns fractions based on the amount of widgets that are visible. This can't be done if you have 4 widgets in a row all being partially clipped.
I've changed the ExactNumWidgetsOnScreen to ExactNumRowsOnScreen and adjusted the behaviors accordingly. This fixed an issue where if there wasn't enough items to fill a row it was assumed that we weren't at the start of the list.

So the end result was that if we aren't at the start and the end of the list then scrolling was necessary.

[CL 2105685 by Justin Sargent in Main branch]