Fixed the build paths option in the editor
When you turn off update navigation automatically in the editor and use
build paths, navigation generation always seems to fail.
In UNavigationSystem::Build() returns early because if NavDataClasses.Num() always returns 0
as it has not initiated yet. This is done in DoInitialSetup() which is called from SpawnMissingNavigationData()
All that needs to be done is reorder the logic slightly and
navigation generation will work again.

GitHub pull request #221
https://github.com/EpicGames/UnrealEngine/pull/221

reviewed by Lukasz.Furman

[CL 2108996 by Thomas Sarkanen in Main branch]