Fixed a crash caused by matinee transaction buffers not using the correct implementation of Begin
I've moved the implementation of UTransBuffer::Begin to a templated function that takes the transaction type to avoid issues like this arising in future.

Also corrected the return value of UTransBuffer::End which was returning the wrong value. This isn't used anywhere in the UE4 codebase but I've fixed it for correctness.
This addresses TTP#338312 - CRITICAL: Regression: EDITOR: MATINEE: CRASH: Occurs when clicking the Start/Stop record button

Reviewed by Gareth Martin, Marc Audy

[CL 2109292 by Andrew Rodham in Main branch]