Auto-save improvements and performance tweaks
https://answers.unrealengine.com/questions/53990/cant-find-file-for-asset.html

This change restores the behaviour of the package auto-saver that was removed in CL# 1950261 due to changes in the way UPackage::PackageDirtyStateUpdatedEvent was called (it was changed to be called only when the package became dirty state was actually changed).

There is now another callback, UPackage::FOnPackageMarkedDirty, which is always called when a package is marked as dirty, even if the package is already dirty. This allows the package auto-saver to watch for changes in-between auto-save periods, rather than be forced to assume that everything that was dirty had become dirty during the last auto-save period. FOnPackageDirtyStateUpdated and UPackage::PackageDirtyStateUpdateEvent have been renamed to FOnPackageDirtyStateChanged and UPackage::PackageDirtyStateChangedEvent respectively.

Additionally, you're now able to restore asset files that haven't yet been saved to disk as a .uasset file. This can happen for newly added or imported assets, as UE4 doesn't generate the associated .uasset file until you do something which causes dirty packages to be saved.

ReviewedBy Thomas.Sarkanen, Max.Preussner

[CL 2109337 by Jamie Dale in Main branch]