Settings editor no longer relies on read-only state of ini files.
The settings editor now queries the appropriate source control state of files that it its editing, rather than using the read-only state (which is only applicable to Perforce).
Also added a throbber to indicate that a source control query is in progress.

TTP# 337061 - EDITOR: SCC: SETTINGS: Settings Editor may not work correctly with some SCC providers

reviewed by Andrew.Brown

[CL 2116519 by Thomas Sarkanen in Main branch]