Added low FPS detection to the editor
FPerformanceWarningNotification constantly monitors the current framerate of the engine and prompts the user when the framerate falls below a desirable value. The heuristics for "low-fps" can be configured with the following CVars:
    * PerfWarn.FineSampleTime, PerfWarn.FineMinFPS, PerfWarn.FinePercentThreshold
    * PerfWarn.CoarseSampleTime, PerfWarn.CoarseMinFPS, PerfWarn.CoarsePercentThreshold

These variables define the length of sample (in seconds), minimum framerate, and percentage threshold that the sample needs to have been under the minimum framerate to prompt the user. Auto-application of scalability settings will be adde soon.

This addresses TTP#337196 - UE4: EDITOR: Low fps detection, Inform User and provide solution using lower scalability settings

Reviewed by Tom Sarkanen, Matt Kuhlenschmidt

[CL 2117735 by Andrew Rodham in Main branch]