Prevented behavior tree nodes from going pink after undo
Made sure node owner is up-to-date after undo. This fixes the outer of the node being incorrect post-undo. Nodes are renamed away non-transactionally when deleted and cleaned up, so the change to the outer gets ignored and is not restored normally. This ensures that it gets properly fixed up.

TTP# 339299 - UE4: BEHAVIORTREE: Node turns pink from Undoing a delete

reviewed by Lukasz.Furman

[CL 2117819 by Thomas Sarkanen in Main branch]