Editor performance monitor can now disable realtime viewports and automatically apply scalability settings
When the monitor detects that the editor is running under an acceptible framerate it will set a flag which causes a benchmark to run on start-up, choosing the most optimal engine display settings for the system. This can be disabled by manually tweaking the settings.
The feature can also be turned off entirely in the editor preferences under Level Editor/Miscellaneous.

Also fixed a shutdown crash by making the performance monitor a member of UUnrealEdEngine. I've had to move some files around to make this possible.

Reviewed by Matt Kuhlenschmidt

[CL 2118280 by Andrew Rodham in Main branch]