Fixing inconsistent expansion of input settings properties
TTP#335353 - EDITOR: Event and axis mapping group expansion behavior is unintuitive

Expansion state was based on the name of the action/axis mapping so when one was renamed it had no state for the new name. Added code to record the state of expansion before renaming and reset it once the mapping groups have been recreated. This required a new function adding to the IDetailGroup class so that I could record its expansion state and I also needed to be sure I only recorded the state when the commit callback actually changed something.

[CL 2119224 by Matthew Griffin in Main branch]