Implemented multi-line editable text IME support
TTP# 336464 - Editor: Finish the Multiline Editable Text Block

This is essentially the same as the version for SEditableText, but it also has to deal with converting the FTextLocation points used within a text layout, into a int32 as used by the IME interface (as the IME always works with the document as a single flat string).

To make this easier, there's now a FTextLayout::FTextOffsetLocations helper class which provides functions to let you convert between the two coordiante spaces. A FTextOffsetLocations can either be generated as an optional parameter to FTextLayout::GetAsText, or, via GetTextOffsetLocations if you don't need the string itself.

This also fixes some general IME issues that were affecting both the multi-line and single-line editable text:
 1) TSF based IMEs no longer allow you to move the cursor outside of the composition range (this is consistent with how the other IME backends work).
 2) IMM based IMEs will now remove the composition text from the document if you choose to cancel your composition.
 3) All IMEs will now finish their current composition and start a new one if you move the cursor via the mouse.

ReviewedBy Justin.Sargent

[CL 2119729 by Jamie Dale in Main branch]