Use 32-bit thread ids on Linux.
This is less portable approach than pthread_self(), but the latter requires more intrusive changes throughout the engine (upping thread ids to 64 bit or even to an opaque type), which will break NetworkProfiler serialization and don't feel warranted in general.

Thanks to x414e54 for bringing this up on #ue4linux.

#codereview Josh.Adams, Daniel.Lamb, JJ.Hoesing

[CL 2122755 by Dmitry Rekman in Main branch]