Asset editors docked into the MainFrame area will now correctly restore to that area when next re-opened
TTP# 338363 - When I relaunch the editor it reopens the items in the wrong place.

IÆve added a "DockedToolkit" tab which resides next to the "LevelEditor" tab. This is used for two purposes; firstly, it can be used to identify asset editors that have been dragged into the MainFrame, and secondly, it's used as a marker (akin to the "StandaloneToolkit" tab) to restore the asset editor to the correct place.

When an asset editor is closed, an entry is added to the GEditorUserSettingsIni file so that FAssetEditorToolkit::InitAssetEditor can restore the asset editor to either the "DockedToolkit" or "StandaloneToolkit" area.

ReviewedBy Thomas.Sarkanen
#codereview Max.Preussner

[CL 2124274 by Jamie Dale in Main branch]