Editor Preferences - Fix an issue with user default array serialization which caused duplicated values in the UI.
TTP# 340060

There is still an issue with array serialization where values are repeated, e.g. Level Editor->Viewports->Decimal Grid Intervals.  User defaults are not saved with any index information so when when they are deserialized the last instances of the changed values are swapped, instead of the ones that the user actually changed.  Fixing that seemed like it would take a larger refactor to the user default serialization and seemed outside of the scope of this bug fix.

[CL 2176373 by Frank Fella in Main branch]