Fix branching creating multiple files
Branching of files needed to be removed (by Bob Tellez) as it was causing crashes and mutiple duplicate files to show up in the content browser. This fix re-instates working branches in the Editor.
The idea is that we now do the copy/duplicate/rename operations first, then perform the 'branch' once the files are finished with by the Editor. This keeps the asset registry & directory watcher systems happy & leaves the issue of branching to source control alone.
Because of the way SVN copy works, the SVN verison of this is slightly icky. SVN copy does not allow a branch-copy over an existing file in the workspace (even if not under source control), so we have to move the file into a temp directory, do the copy, then re-move the file back over the top of its old location.

TTP# 334923 - EDITOR: Perforce Integration (Move -> Delete + Add instead of Integrate)

reviewed by Max.Preussner,Bob.Tellez,Matt.Kuhlenschmidt

[CL 2180124 by Thomas Sarkanen in Main branch]