Allowing different sample rates for Opus encoding
Choose an Opus compatible sample rate based on the input data and store that rate with the audio so that we know what to decode at.
Also changed the bit rate calculation so that it's more closely related to the original sample rate and number of channels, with the quality determining overall size reduction.

[CL 2223219 by Matthew Griffin in Main branch]