Fixed FTransform details customization not dirtying packages correctly.
Made sure to call NotifyFinishedChangingProperties so that packages are dirtied when editing values.
Also increased precision in floating-point SetValue. We should not go via strings to do this, but we dont have any other way at the moment, so this is the next best thing.

TTP# 341202 - BLUEPRINT:  A transforms default value does not get carried over from the set values in its structure

reviewed by Andrew.Brown

[CL 2228482 by Thomas Sarkanen in Main branch]