Fixes Interactive Tool Tips not being interactive any more.
This TTP is not complete; only did safe step because of 4.4 deadline.
Further cleanup is required. The function SWindow::AcceptsInput() should include the call to IsWindowHousingInteractiveTooltip() and all other calls to IsWindowHousingInteractiveTooltip() should be deleted. However, the implications of that change must be studied. For the moment, IsWindowHousingInteractiveTooltip() is being called strategically alongside AcceptsInput() in the two locations that minimize potential fallout.

[CL 2231202 by Nick Atamas in Main branch]