Added FText::IdenticalTo and made SRichTextBlock take its text as an attribute
FText::IdenticalTo compares the string pointers inside the FTexts to see if they're identical. This can be used as an early out when checking if a cached copy of an FText is the same as the current version of an FText (since if the string pointers are identical, then the strings themselves shouldn't have changed).

SMultiLineEditableText::Tick has been changed to use FText::IdenticalTo to speed up its equality testing when using a bound FText attribute.

SRichTextBlock has also been updated to take its text as an attribute. This wasn't done originally due to the performance considerations when caching and comparing the FText, but FText::IdenticalTo alleviates those concerns.

ReviewedBy Justin.Sargent

[CL 2235719 by Jamie Dale in Main branch]