SScrollBox - Add support for horizontal scrolling to the core widget and UMG.
There is more copy/paste here than I wanted.  I had some unified layout/construct code, but with all of the templates and additional function calls, it was much harder to follow, and performance would have probably taken a hit, so I opted for the clearer duplicated code in a few spots.

[CL 2235776 by Frank Fella in Main branch]