Fixed stall when landing on Foliage Actors.
We now check the primary actor bCanEverTick to fix issues with landing on Foliage actors with thousands of components. We need to depend on other component's Ticks in the general case to deal with things like Timelines driving the position of moving platforms etc.
This will prevent dependencies from being created for Blueprints that use timelines but do not implement a Tick() event, but this is a fairly simple workaround and the side effect is minor (a one frame lag in movement between base and character).

TTP# 341962 - EDITOR: FOLIAGE: Editor freezes temporarily when jumping on foliage with collision in user level

#codereview Jack.Porter,James.Golding

[CL 2244275 by Thomas Sarkanen in Main branch]