Fixed some centering and scaling issues with wrapped Canvas text
TTP# 343372 - CANVAS: WrappedPrint() is not functioning correctly when used with centering and scaling

Updated some places in the Canvas string rendering that were using the older StringSize function (which has a hard-coded scale of 1) rather than the newer UCanvas::CanvasStringSize (which correctly measures using the current scale).

The centering on Y was trying to center each line on its own height, rather than center the entire block of text on its total height. This has been updated to apply a centering based upon the total height of the text being drawn.

Fixed some places that were doing float -> int32 -> float truncation when measuring.

#codereview Thomas.Sarkanen, Max.Preussner

[CL 2256144 by Jamie Dale in Main branch]