[AUTOMERGE]
Fix for TTP 343992: Server crash due to AI behavior tree stopping during AI's destruction in some specific circumstances.

NOTE: We still need to track down exactly why we're trying to stop the behavior tree when it's in the process of being destroyed, as well as why this would only be a problem in very specific circumstances.  However, this fix should in the meantime go ahead and prevent at least the specific crash in question.

Also, if the crash would have occurred, it now prints some information which may be helpful in tracking down exactly what is happening.

(Also, moved call to Super function (which is empty anyway) to the top of the function since it doesn't need to occur between other pieces of functionality in BTService_BlueprintBase.)

#Fortnite

--------
Integrated using branch Ue4-To-UE4-Fortnite-Simple (reversed) of change#2258242 by Daniel.Broder on 2014/08/15 14:43:50.

[CL 2258255 by Daniel Broder in Main branch]