Fixes to list view overscroll
Inertial scroll was always being cleared when scrolling in the overscroll region. This meant that scrolling out of the region did not correctly sample.
    Solution: ClearInertialScrollVelocity on mouse/touch up (if overscrolling).
Side effect: Inertial scroll still operates even when overscrolling(this was previously also the case, but they were cleared on tick).
    Solution: Add CanUseInertialScroll to control when inertial scroll can occur

This addresses TTP#339241 - iOS style scrolling has odd velocity and drag properties when dragging/velocity past the end of the list

Reviewed by Nick Atamas

[CL 2265878 by Andrew Rodham in Main branch]