Improved the validation for working out if a path is too long to cook
TTP# 332328 - Content Browser: Need new warnings about long relative path lengths for assets.

https://answers.unrealengine.com/questions/60187/the-path-to-the-asset-is-too-long-for-cooking-im-l.html

The previous logic assumed that the project would always be in the same source tree as UE4, which isn't the case for content creation and Rocket projects.

In the case where the project was outside the UE4 source tree, the validation often failed as the path was never trimmed to remove the relative part of it (between the project and the UE4 binary).

This change gracefully handles both cases, and now performs two sets of validation:
 1) It checks that the content can be cooked based on where the project actually is (for local cooking)
 2) If it's an internal build, it also checks that the content will be cookable via our build machines (forthe build farm cooking)

ReviewedBy Andrew.Rodham
#codereview Max.Preussner

[CL 2272171 by Jamie Dale in Main branch]